+++ database/migrations/2019_08_01_120630_create_classrooms_table.php
@@ -8,7 +8,6 @@ class CreateClassroomsTable extends Migration
 {
     /**
      * Run the migrations.
-     *
      * @return void
      */
     public function up()
@@ -26,23 +25,25 @@ class CreateClassroomsTable extends Migration
                 ->onDelete('cascade');
         });
 
-        Schema::create('classroom_translations', function(Blueprint $table)
-        {
+        Schema::create('classroom_translations', function (Blueprint $table) {
             $table->increments('id');
             $table->string('category')->nullable();
             $table->json('label')->nullable();
             $table->string('locale')->index();
 
-            $table->integer('classrooms_id')->unsigned();
-            $table->foreign('classrooms_id')->references('id')->on('classrooms')->onUpdate('cascade')->onDelete('cascade');
+            $table->bigInteger('classrooms_id')->unsigned();
+            $table->foreign('classrooms_id')
+                ->references('id')
+                ->on('classrooms')
+                ->onUpdate('cascade')
+                ->onDelete('cascade');
 
-            $table->unique(['classrooms_id','locale']);
+            $table->unique(['classrooms_id', 'locale']);
         });
     }
 
     /**
      * Reverse the migrations.
-     *
      * @return void
      */
     public function down()
